change values from int to float64

+++ api.go
@@ -1,8 +1,8 @@
 package main
 
 type SwitchRequest struct {
-    Device string `json:"device"`
-    Command string `json:"command"`
-    Value int `json:"value"`
-    Duration int `json:"duration"`
+	Device   string  `json:"device"`
+	Command  string  `json:"command"`
+	Value    float64 `json:"value"`
+	Duration int     `json:"duration"`
 }
